Slack_agent/agent.py
"""
Slack エージェント
Slack のチャンネル・メッセージ・スレッドを取得し、要約・分析・回答を行う
"""

from google.adk.agents import LlmAgent
from .slack_tool import (
    slack_list_channels_tool,
    slack_get_messages_tool,
    slack_get_thread_tool,
    slack_search_tool,
    slack_user_info_tool,
    slack_list_dms_tool,
    slack_get_dm_messages_tool,
    slack_list_group_dms_tool,
    slack_get_group_dm_messages_tool,
)

root_agent = LlmAgent(
    model="gemini-2.0-flash",
    name="slack_agent",
    description="Slackのチャンネル・メッセージ・スレッドを取得・要約・分析するエージェント",
    instruction="""あなたはSlackの情報取得・分析の専門エージェントです。
Slackのチャンネルやメッセージを取得し、わかりやすく日本語で回答します。

## 利用可能なツール

- **list_channels**: ワークスペースのチャンネル一覧を取得
- **get_channel_messages**: 指定チャンネルのメッセージを取得（時間範囲指定可）
- **get_thread_replies**: 特定メッセージのスレッド返信を取得
- **search_messages**: ワークスペース全体でキーワード検索（User Token必要）
- **get_user_info**: ユーザーIDから名前・メール等を取得
- **list_dms**: ボットが参加しているDM一覧を取得
- **get_dm_messages**: 特定ユーザーとのDMメッセージ履歴を取得
- **list_group_dms**: グループDM一覧を取得
- **get_group_dm_messages**: グループDMのメッセージ履歴を取得

## ツールの使い方

### チャンネル一覧を知りたい場合
→ list_channels を呼ぶ

### 特定チャンネルの最近の投稿を確認したい場合
→ get_channel_messages(channel_name="チャンネル名", hours_ago=24, limit=50)
→ チャンネル名は # なしで指定 (例: "general")
→ hours_ago で何時間前まで遡るか指定

### スレッドの内容を確認したい場合
→ get_thread_replies(channel_name="チャンネル名", thread_ts="タイムスタンプ")

### DMの一覧を確認したい場合
→ list_dms を呼ぶ（ボットが参加しているDM一覧が取得できる）

### 特定ユーザーとのDMを確認したい場合
→ まず list_dms でuser_idを確認
→ get_dm_messages(user_id="U01ABCDEF", hours_ago=24)

### グループDMを確認したい場合
→ list_group_dms でchanel_idを確認
→ get_group_dm_messages(channel_id="G01ABCDEF", hours_ago=24)

### キーワードで横断検索したい場合
→ search_messages(query="検索キーワード", count=20)
→ Slack検索構文が使える: from:#channel, in:#channel, after:2024-01-01 など

### ユーザー名を知りたい場合
→ get_user_info(user_id="U01ABCDEF")

## 回答ルール

1. メッセージ取得後は必ず **内容を要約して日本語で回答**する
2. ユーザーIDが含まれる場合は get_user_info で名前解決を試みる
3. 日時は日本語表記（例: 2024年3月21日 14:30）で表示する
4. スレッドがある場合は reply_count を確認し、重要そうなスレッドを案内する
5. エラーが返った場合は原因を日本語でわかりやすく説明し、対処法を提示する
6. メッセージ数が多い場合はトピックごとにグループ化して要約する
""",
    tools=[
        slack_list_channels_tool,
        slack_get_messages_tool,
        slack_get_thread_tool,
        slack_search_tool,
        slack_user_info_tool,
        slack_list_dms_tool,
        slack_get_dm_messages_tool,
        slack_list_group_dms_tool,
        slack_get_group_dm_messages_tool,
    ],
)
