Slack_agent/pyproject.toml
[project]
name = "slack-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.23.0",
    "slack-sdk>=3.27.0",
]
